+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.30.1] - 2025-01-10
+
+### Added
+
+- Added `_load_page` method to `TimeseriesCollection` to easily allow individual pages, allowing to not do auto-pagination for use cases where it is not needed.
+- Added docstrings to the dataset client constructors.
+
 ## [0.30.0] - 2024-11-29
 
 ### Changed
@@ -91,7 +98,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Released packages: `tilebox-datasets`, `tilebox-workflows`, `tilebox-storage`, `tilebox-grpc`
 
 
-[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.30.0...HEAD
+[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.30.1...HEAD
+[0.30.1]: https://github.com/tilebox/tilebox-python/compare/v0.30.0...v0.30.1
 [0.30.0]: https://github.com/tilebox/tilebox-python/compare/v0.29.0...v0.30.0
 [0.29.0]: https://github.com/tilebox/tilebox-python/compare/v0.28.0...v0.29.0
 [0.28.0]: https://github.com/tilebox/tilebox-python/compare/v0.27.2...v0.28.0